Set up skeleton for zoo example.
