+++ example/tutorial/.gitignore
@@ -58,3 +58,4 @@ typings/
 .env
 
 zoo-data/*
+*_params
