--- example/tutorial/common.js
@@ -1,3 +1,0 @@
-exports.dataRoot = './zoo-data/'
-exports.imageRoot = exports.dataRoot + 'images/';
-exports.databasePath = exports.dataRoot + 'zoo.sqlite';
