--- example/tutorial/form.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <link rel="stylesheet" href="styles.css">
-</head>
-<body>
-  <div style="max-width:500px">
-<form action="http://localhost:3000/upload/animal" method="post" enctype="multipart/form-data">
-<ul class="form-style-1">
-    <li><label>Animal <span class="required">*</span></label><input type="text" name="animal" class="field-long" /></li>
-    <li><label>Name <span class="required">*</span></label><input type="text" name="name" class="field-long" /></li>
-    <li>
-        <label>Habitat <span class="required">*</span></label>
-        <input type="text" name="habitat" class="field-long" />
-    </li>
-    <li>
-        <label>Profile Pic <span class="required">*</span></label>
-        <input type="file" name="profile_pic"/>
-    </li>
-    <li>
-        <label>Description <span class="required">*</span></label>
-        <textarea name="description" id="description" class="field-long field-textarea"></textarea>
-    </li>
-    <li>
-        <input type="submit" value="Submit" />
-    </li>
-</ul>
-</form>
-  </div>
-</body>
-</html>
