+++ example/tutorial/nginx/.gitignore
@@ -1,0 +1,3 @@
+*_params
+*_temp
+logs/*
