+++ example/tutorial/nginx/nginx.conf
@@ -1,0 +1,34 @@
+events {}
+
+http {
+  include lightstep_access_token_params;
+
+  upstream backend {
+    include upstream_params;
+  }
+
+  server {
+    listen 8080;
+    server_name localhost;
+
+    location = / {
+      proxy_pass http://backend;
+    }
+
+    location = /animal {
+      proxy_pass http://backend;
+    }
+
+    location = /upload/animal {
+      proxy_pass http://backend;
+    }
+
+    location / {
+      root www; 
+    }
+
+    location ~ \.jpg$ {
+      include image_params;
+    }
+  }
+}
