+++ example/tutorial/nginx/www/add.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>New Admittance</title>
+</head>
+<body>
+<form action="/upload/animal" method="post" enctype="multipart/form-data">
+  Name:<br>
+  <input type="text" name="name"><br>
+  Animal:<br>
+  <input type="text" name="animal"><br>
+  Profile Pic:<br>
+  <input type="file" name="profile_pic"><br><br>
+  <input type="submit" value="Submit">
+</form>
+</body>
+</html>
