+++ example/tutorial/node/common.js
@@ -1,0 +1,1 @@
+exports.databaseName = 'zoo.sqlite';
