+++ example/tutorial/node/views/animal.pug
