+++ example/tutorial/node/views/index.pug
@@ -11,4 +11,6 @@ html
                 img(src=animal.thumbnail_pic)
               br
               | #{animal.name}
-
+    a(href="/add.html")
+      | New Admittance
+
