+++ example/tutorial/stop.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+
+DATA_ROOT=$PWD/zoo-data
+
+while read pid; do
+  kill $pid
+done <$DATA_ROOT/backend_pids
+rm $DATA_ROOT/backend_pids
+
+nginx -p nginx/ -c nginx.conf -s stop
